chore #13 : RichText BindingAdapter 적용 및 OBJ 이슈 해결
