+++ presentation/src/main/res/layout/view_rich_view_type.xml
@@ -22,6 +22,7 @@
         <TextView
             android:id="@+id/text_rich"
             android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
+            android:layout_height="wrap_content"
+            richText="@{richViewType.richText}"/>
     </LinearLayout>
 </layout>
